pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-socketio>=5.5.1",
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "spacy>=3.8.5",
    "sqlalchemy>=2.0.41",
    "websockets>=15.0.1",
    "requests>=2.32.3",
    "stripe>=12.3.0",
    "redis>=5.0.1",
    "PyJWT>=2.8.0",
]
